Add new sounds with sound manager which is awesome to use
