+++ Assets/Scripts/Managers/AudioManager.cs
@@ -1,0 +1,46 @@
+﻿using UnityEngine;
+using System;
+using System.Collections.Generic;
+
+public class AudioManager : MonoBehaviour
+{
+    public Sound[] sounds;
+
+    public static AudioManager instance;
+
+
+
+    private void Awake()
+    {
+        DontDestroyOnLoad(gameObject);
+        if (instance == null)
+        {
+            instance = this;
+        }
+
+            foreach (Sound sound in sounds)
+        {
+            sound.audioSource = gameObject.AddComponent<AudioSource>();
+            sound.audioSource.clip = sound.clip;
+
+            sound.audioSource.volume = sound.volume;
+            sound.audioSource.loop = sound.loop;
+        }
+    }
+
+    private void Start()
+    {
+        PlaySound("BackgroundSound");
+    }
+    public void PlaySound(string soundName)
+    {
+        Sound s = Array.Find(sounds, sound => sound.name == soundName);
+        if(s == null)
+        {
+            Debug.Log("Wrong " +s+ " sound name in editor!");
+            return;
+        }
+
+        s.audioSource.Play();
+    }
+}
