+++ Assets/Scripts/Managers/TimeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d62969a6be9b214b9770999ade4a0ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
