--- Assets/Scripts/Sound/AudioManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cae4709e00fa2624392d30a490a4c94a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
